Disable `cargo about` talking to clearlydefined.io.
This is a workaround for a lack of rate limiting causing failures
(<https://github.com/EmbarkStudios/cargo-about/issues/218>), but also,
it does not seem to be necessary for my use case (no unknown licenses
are reported), so we might as well *not* talk to an external service.

Also add `--fail` option to fail the build if there ever is a problem
license.