test-renderers: use `UNALTERED_COLORS` as the base graphics options.
This means that introducing new effects in the default set will not
break the existing tests. (It does also mean that we lose coverage of
the interactions between various color-affecting options, but we can
always add regression tests as needed, and we never had coverage of
multiple non-default options.)

The antialiasing test images required updating, because it no longer
includes fog or lighting. Note that the wgpu image is incorrect as it
does not contain antialiasing within block surfaces.