Move mesh code to new crate `all-is-cubes-mesh`.
It has a (mostly) clean boundary, and it serves a very specific purpose,
so it makes a good candidate for splitting. This doesn't save very much
compile time on changes to all-is-cubes, because -mesh is relatively
tiny, but it saves many seconds for changes to -mesh itself. It also
removes the `fnv` and `bitvec` dependencies from `all-is-cubes`.