port: Add glTF export (currently `BlockDef`s only).
This is a small step towards merging the code for “record” and “export”,
or at least not having two completely different architectures: now we
have a working example of exporting glTF, so we can incrementally move
the recording system onto export infrastructure. Additionally, this may
be useful for actually exporting block definitions.