desktop: Fix export-as-record; add test.
I must have broken this during some refactoring (...or it never worked?)
and it would emit the “must be called from a Tokio runtime” error
instead of exporting anything. Fixed that by passing the `Handle`
explicitly. Of course, the right answer is to refactor and stop using
`block_on()` in this fashion.