port: add `ExportSet::member_export_path` for nonconflicting file names.
Just enough to make a file-per-member export work at all; not a complete
solution yet because it doesn't: sanitize text of the member names (e.g.
replacing `/`) or avoid duplicates after doing that. Eventually we will
want to make this done for the _entire_ ExportSet at once to fix these
problems.