test-renderers: Implement/stub out test options for `nextest`.
<https://nexte.st/> expects each test binary to implement these options
like the default test harness does.

Note: It is still not _wise_ to run `test-renderers` using
`cargo nextest`, because it wastes a lot of time on GPU and `Universe`
initialization by running each test function in a separate process,
but this makes it _possible_, which may be useful.

If we decide to do any more compatibility, we might want to use the args
parsing from <https://docs.rs/libtest-mimic/> instead of writing our
own. Unfortunately, it doesn't offer filtering without running.