xtask: Pass `--no-deps` to `cargo doc`.
I saw an error in CI where it ran out of disk space during documenting
(on a Windows runner). Hopefully that is just a flake, but regardless,
it made me realize that we have no real need to be spending time on
building dependencies' docs in CI since that doesn't give us any useful
warnings.