mesh: Make `get_block_mesh()` non-`Option`-al.
This simplifies the interface so that `SpaceMesh` does not need to
specifically have its own error handling; now the `GetBlockMesh`
implementation gets to choose what to do about errors. Currently,
the same behavior is preserved, by returning the new constant
`BlockMesh::EMPTY_REF`. In the future, we might instead decide to
panic (on the premise that this really shouldn't happen) or display an
error marker block (for robustness without hiding the fact that
something went wrong).