rerun: Add unused `rerun` dep and feature.
These will be used to implement external logging/visualization of
physics behavior, which I hope will finally squash some physics bugs.
The advantage of using Rerun over internal visualization is that it
provides replay, a camera independent of the game camera, and a tree
inspector of the logged data. I could build all of these things in-game,
but that would be a much longer path. Once the bugs are fixed I can
consider whether this support should be deleted.