Hook up deserialized `URef`s.
Lot of kludges here. Most notably, `Universe` and `URef` share a
thread-local variable to enable creation of the `URef` structures.
The alternative would have been to modify `URef` so that it can have
its references swapped out during a `visit_refs()`, i.e. make them
interior mutable, which I judged to be a worse option. In the future
we may want to look into serde's `DeserializeSeed` to do this in a
cleaner fashion.