Use `DynListener` inside `Block` and `Modifier`'s `listen_impl()`s.
This should save some compile time by keeping things non-generic, and
avoid boxing the `Listener` (`.erased()`) for _each_ dependency the
block has, at the cost of boxing it even if it is not installed in any
notifiers. (We could do even better by making the boxing lazy, but
that doesn't immediately seem worthwhile; *this* change is a definite
simplification.)

I also want to do some further refactoring which will be easier when
the listening logic has no generics.

Running `gen_bench` (which does a lot of setting new blocks and thus
triggering listening) shows a small uniform improvement, but that might
be just noise.