Give `Primitive::Atom` a struct for its data.
This will make it more convenient to add further fields. I plan to
immediately add a `collision` field (replacing
`BlockAttributes::collision`), and in the future there may be other
material properties.

The rationale for defining a struct instead of a struct-style enum
variant is that editors and builders may benefit from having an `Atom`
type, and if I decide to make the `Atom` struct non-exhaustive then
it will be more convenient than a non-exhaustive variant.