Move `Space` index management into new module `space::palette`.
The palette data is now in a single private `struct Palette`, which may
be constructed or replaced as a whole, which is useful for `Space`
construction and `fill_uniform()`, and hopefully clarifies the code
(though the exact consistency requirements between `Space` and `Palette`
are not yet documented).

The motivation for this change is to simplify the implementation of a
new function `fill_with_palette()` to more efficiently implement
deserialization, imports, and any other situations where we are starting
with a palette (though not a `Palette` since that's an implementation
detail).

Future work: We might want to rename `SpaceBlockData` to `PaletteEntry`
and `Space::block_data()` to `Space::palette()` or something like that,
not because the outside must match the inside but because my additional
experience with voxel systems says that “palette” is a common term for
this concept.