Begin attribute merging in `Modifier::Composite`.
Currently only the `selectable` attribute is merged.

Note that the correct behavior is not at all clear; arguably, the
selectable attribute should be turned off if it is turned off in
either case because non-selectability is a special case. However, the
primary example we have in front of us is that if we want to to compose
an `AIR`-like block and something else, the something else's
selectability should be preserved.