wasm: Add a test for GPU renderer on Wasm.
The motivation for this test is that I just had to add some Wasm-specific
code to the implementation, but it was never actually used, so therefore
never tested.

In the future we could consider making the entirety of `test-renderers`
run in a browser, but that will require additional harness work.