Implement serialization for `BehaviorSet`.
Unfortunately, there are no `Behavior`s right now that are both suitable
to serialize and defined inside the `all_is_cubes` crate. So, any actual
use (or full testing) of the system will need either such behaviors, or
implementing extensible serialization (which is a tricky matter since it
implies some kind of extension registry — but we will probably need it).