CI: Rename `aic.yml` to `ci.yml`.
This aligns with other projects' conventional arrangement. At the time
I was doing the initial setup, I wasn't clear on what the role of this
file would be; now it's clearer that it makes sense that "this is the CI
file", meaning that this is the stuff that runs on every PR or push.