Give universes an inherent choice of time step.
This will allow discrete simulations, such as blocks interacting with
each other and updating themselves, to proceed at a rate designed to
suit the universe, without requiring a hardcoded constant time step.

This commit adds the types `TickSchedule` (defining the tick length)
and `Clock` (generating `Tick`s), and removes the fixed tick length
from `FrameClock`. However, the goal is not yet achieved:

* The schedule cannot yet be changed.
* The schedule is not yet serialized.
* Also, `VuiManager` needs to arrange for itself to be able to tick on
  a possibly different schedule; right now it will use the game ticks
  even though it shouldn't really.

The `Clock` also introduces a concept of “phase”, where there is a
per-tick counter that loops over a short period. This will be used
to allow simulations to run on an intentionally slow schedule (e.g.
a block that changes state only 5 times per second while the tick
rate is 60) without needing independently running timers. However,
that is not yet propagated through `Tick`s.