CI: Add a bunch of `df` commands to keep track of disk space usage.
We hit runner limits, and based on what I can see from the first
iteration of this debugging, I *think* this is just dependency
updates causing lots of stale data mixing with fresh data. I'm going to
try deleting the caches to effectively expire the old data faster.
However, we are probably pretty close to limits overall, so it would
be good to think about any possibilities of trimming dependencies.