Add `Gridgid` transform type.
This transform is stricter than `GridMatrix` in that it does not allow
scaling or skewing; it is always a rigid transformation, and always
invertible. This turns out to be what we actually want in most cases,
so switching to it should remove various `unwrap()`s and edge cases.

Also added `GridRotation::rotate_vector()`, which is part of the
implementation and will be generally useful.

Future work will switch many uses of `GridMatrix` to `Gridgid`.