desktop: Do universe building and such inside `inner_main()`.
This is a small step towards being able to do the tasks like creating a
new universe any time, not just at the beginning.

Note that we're still not yet actually starting the event loop until
afterward, so the window does not appear during this phase (at least on
macOS).