mesh: introduce a `Plane` trait for 2D slices of textures.
Instead of `Tile` directly providing texture coordinates, a `Plane`
must be created first, which narrows down the scope to a particular 2D
slice through the 3D data. Thus the allocator has an opportunity to use
different coordinates for different planes, rather than requiring a
unified 3D space.

This will allow texture allocator implementations to work with 2D
textures instead of 3D, if that suits them -- such as the glTF
exporter.