glTF: compose texture atlas.
This code is definitely incorrect but it's hard to test well, because
we do not yet have the necessary texture coordinate remapping to *use*
these textures. That will be later work.

The choice of `rectangle-pack` as a library was based on its popularity,
its lack of dependencies, and its support for 3D packing (which I might
try using for our 3D texture allocator in place of our own `Alloctree`).
I have not yet tried other packers -- `etagere`, `guillotiere`,
`crunch`, and `packos` are on the list of alternatives to investigate
if I have reason to.