Compute `EvaluatedBlock::color` via raytracing.
This computes a more accurate color for the exterior of a block,
disregarding interior voxels that are trivially hidden by the outside
surfaces. It does not produce a true average color from all viewing
angles, since only axis-aligned rays are considered, but it should
produce good results for typical content.

Note that render tests' output changes because the color values are
used in light transport calculations.

A further improvement might be to split the color into six separate
face colors. However, that's fraught because it suggests doing the same
to `Evoxel`, which is a can of worms for the volumetric interpretation.