glTF: separate color and texture coordinates.
It would be optimal to have two different vertex formats for colored
and textured vertices, but this will be easier to make progress with,
as `all-is-cubes-mesh` has no support for splitting up meshes by
material other than transparency, so we would have to do it as another
postprocessing step.

This commit offers no practical benefits but is a step towards
actually working glTF textures.