Make serialization support optional.
This will improve compilation time for anyone using `all-is-cubes` as a
library (which is a use case I want to support, perhaps foolishly) in an
application where serialization is irrelevant.

It is also a small step towards hopefully being able to support `no_std`
compilation of `all-is-cubes`; if this works at all then we will also
want serialization, but that will require replacing `flate2` so it can
be a followup project.