Introduce `Instant` trait and `Deadline` type to abstract over clock sources.
* New trait `Instant` replaces using the `instant::Instant` shim within
  `all-is-cubes`.
* New struct `Deadline` wraps an `Instant` and adds values for the
  special cases of “do as little work as possible” and “finish the job,
  however long it takes” without choosing arbitrary time values.
* Other crates than `all-is-cubes` have not yet been made generic.

This is a small step towards hopefully being able to support `no_std`
compilation of `all-is-cubes`. It also means we're not relying on the
`instant` library in particular for compatibility.

I'm not a fan of how this requires making some significant code generic,
but I haven't thought of a way to solve that at a reasonable cost
without reaching for more esoteric options like stack-allocated
limited-size `dyn`s.