Use `rust-toolchain.toml` to specify target and component requirements.
This way, the requirements are automatically available to all users,
not just specifically the GitHub Actions CI.

A caveat is that `rustup override set` ignores the file entirely, so
in order to change toolchains we have to patch the file instead of
using the override command. Still seems better overall.