Change `Evoxels` storage to `Arc` so it can be cheaply cloned.
This is part of a larger project to make `EvaluatedBlock` cheap to
clone, so that e.g. meshing can be done asynchronously -- we don't
want to copy the data because the `Evoxels` data can be huge.
(The alternative of `Arc<EvaluatedBlock>` would be inconvenient for
`AIR_EVALUATED`, but on the other hand would mean fewer individual
refcount updates.)