ui: Generalize `set_universe_async()` to “main task”.
The main task is not specifically expected to produce a universe,
but instead can take various actions interleaved with stepping.

This immediately lets us provide error dialogs for failed drag-and-drop
imports, which was missing; I hope to expand the capabilities until it
can provide the functionality of `record_main()` without dedicating a
loop to that, thus enabling simultaneous recording and interaction.