Add `Primitive::Text`.
Empirically, it has turned out that making text rendering consist of
writing blocks into a `Space` is a bad idea. It is slow, awkward to set
up since it requires mutating the `Universe`, and also probably not very
user-editable. Therefore, I'm adding a text rendering primitive. It will
initially support only builtin fonts, but I expect to add fonts as
universe members in the future, and probably migrate away from using
`embedded_graphics` for the rendering.

This is a minimal proof of concept.  Further work is needed make this
usable for all text applications; I expect to add the following:

* Explicitly chosen resolution, so that compositions with other blocks
  know what they're getting in to, and so that thickness of the text is
  well defined.
* Replace the `anchor_point` concept with a bounding box (which the text
  may be permitted to spill out of).