Execute lighting calculations in parallel.
This slightly changes the results due to batching (updates in a batch
don't see other updates in the same batch), but does not introduce
nondeterminism.

Note that this is a performance improvement, but what we should really
be doing is running light calculations in the background all the time,
which will require more work to make the data available for use.