Remove `impl Deref for BlockDef`.
The implementation is more trouble than it’s worth, especially now that
`BlockDef` performs evaluation caching and we don't necessarily want to
call the dereferenced `Block::evaluate()`; and almost all of the deref
usage other than that was explicit `&*` rather than implicit.

Its function is replaced with a `.block()` method.

Future work: get rid of the `.block().evaluate()`s.