desktop: Make `indicatif` and `simplelog` cooperate.
We now hide the progress bars (if any) when logging, so that no extra
copies of the progress bar appear in the log text.

Note: this has the not-intended effect of causing the progress bars to
entirely vanish when finished, rather than leaving their line with
their time and count performance data. However, we can replace that
functionality with explicit logging if we want it.