Add `EvaluatedBlock::face_colors()`.
This additional color information will be useful for light propagation
and placeholder/low-LOD meshes, and is cheap to compute since it is just
subdividing the results of the aggregation we were doing anyway.

I do not currently intend to use it to add per-face colors to `Evoxel`s,
as that would greatly complicate storage, texturing, and transparency.

The glTF test output changed because of floating-point rounding changes:
the new calculation accumulates the six faces separately before
combining them, rather than accumulating all pixels of all faces
in a single `VoxSum` accumulator.