Add alpha weighting to `EvaluatedBlock::color`.
Previously, the color was the mean color of all of the `EvalTrace`s,
even though influence should be weighted by alpha. Now it is.

I'm not fully confident that this is the right arithmetic, because it's
difficult to write a test case that it is itself definitely correct and
thorough, but I'm confident it's better than the previous alternative.