mesh: Fix `ignore_voxels` being ignored.
This is a partial revert of commit f4bd20f6d09cec7b35939a4099e1305353032be5
which introduced the bug; at the time I misread the code and thought
this case was for a different purpose than it actually is.

What remains is the switch from a `match` to `if let` using
`Evoxels::single_voxel()`, which is still desirable since it saves an
indentation level and avoids needing a dummy branch for match
exhaustiveness.