ui: Partial implementation of multi-block buttons.
Not yet suitable for use because it duplicates the labels onto every
block of the button, rather than placing them centered on the button.
Therefore, button layout still always chooses one block. But we've got
the creation of the multi-block button base shapes working, which is an
important step.

The next step needed is for the composition of button and label to
be performed on creation of the widget controller rather than the
widget, once it knows what its size is. Perhaps the widget controllers
will store prepared `SpaceTransaction`s for each state.

We will also probably need some additional `LayoutTree` features to help
give buttons sizes that are neither excessively stretched nor minimal
(e.g. a stack of buttons which is as wide as their widest label).