Allow `Vol<()>` to be compared with `GridAab`.
This makes sense because `Vol<()>` is a `GridAab` with restrictions.
Arguably the ordering (if it is ever not a ZST) should disqualify this
as a lossy comparison, but we can also see the ordering as irrelevant
in the same way that a `HashMap`'s ordering is irrelevant (though by
that argument we should also let `Vol`s be compared with each other
regardless of ordering, which we do not yet).