ui: Use `Label` widget instead of text rendered in a `Space`.
Future work:
There seems to be some extraneous vertical space in the about box
paragraphs, making the box taller and more loosely spaced than it ought
to be. However, I don't want to block the overall text rendering
migration on getting to the bottom of that, and the space is tolerable
if inelegant, so I'm going ahead with this change.