ui: `LayoutTree::Stack` modifies gravity.
I thought it already did this from the beginning, but it didn't, and
that explains some weirdness going on. This still isn't a complete
solution for desired layouts — in particular, for example, if we make
a table, we might have a Y stack of rows each of which is an X stack,
the rows might want Y centered gravity and there's no way to ask for
that.