desktop: Split out a library crate.
This will be used for debugging techniques like opening a window out of
`test-renderers` for interactive viewing. (It is probably not yet
actually usable that way; in this commit I am focusing on making the
split at all, without too much further refactoring.)

It also demonstrates a slight improvement in build time, because binary
crates don't seem to take advantage of build pipelining, so by adding a
desktop library we allow its codegen phase to proceed in parallel with
the codegen for other library crates (notably `-port` and `-content`).