New block evaluation `Budget` system replacing `depth`.
The old idea of just capping recursion depth was too weak, because
it's easy to make a very expensive block without having very much
recursion. So, instead, we have counters that are only decremented
(while still also having a recursion guard).

By putting it inside the `EvalFilter`, we avoid having multiple
recursively passed parameters, and allow the original caller to both
choose the budget and find out how much of it was used.

Future work:
Expose evaluation costs in `EvaluatedBlock` for diagnostic purposes;
specifically so users can avoid running into them unexpectedly, and
so we can write tests of this new logic.