Generalize `BlockProvider` to non-blocks (as `Provider`).
In the long run, I want to be able to make modules that contain various
different universe-member types, not just blocks. That will require more
design, but I expect this will help. For now, I'm doing this because I
want to be able to map from a regular `BlockProvider` to a set of
`BoxStyle`s.