gpu: Implement background polling on Wasm like on non-Wasm.
This will be necessary in future versions of `wgpu` where
`poll(Maintain::Wait)` is an error on Wasm since blocking is generally
prohibited on web. However, due to what cwfitzgerald believes is a bug
in Firefox, `Maintain::Poll` doesn't actually allow progress to be made,
so the net effect of this commit is to just move the `Wait`ing to a
background task, until such time as the bug is fixed.