Move `BlockBuilder::voxels_fn()` body to a non-generic function.
This reduces size of the desktop binary by 1% and will hopefully have
a similar effect on build times.

Note that the bound on the voxel function's return type is now
`Into<Cow<'a, Block>>` instead of `Borrow<Block>`. This still covers
the actual usage, which is either `Block` or `&Block`. (I considered
using just owned `Block` all the time, but that would avoid a branch
on `Cow` variants at the cost of updating `Block` refcounts a lot,
which seems likely to be worse.)