mesh: Rename `MeshLabel` to `MeshId` and declare it reliably unique.
This is intended to be useful for applications where we need to send the
mesh data as a set of objects added/removed (such as glTF and Rerun)
rather than driving the rendering every frame ourselves. However, this
commit only introduces the property and does not make use of it.