Add `Face6::from_snapped_vector()`.
I expect to use this code in some `RotationPlacementRule`s; but I'm
committing it by itself because that work got stalled but I want to
reduce the number of inactive development branches, and it's got a good
chance of coming up useful at *some* point.

(An uninvestigated question: are its edge-case choices consistent with
the choices that `Raycaster` makes for a ray through the center of a
block? It might make sense to make them so.)