Make `gltf-render` test not run by default even when the feature is enabled.
Since the gltf render tests don't pass yet, this is necessary for
`xtask test-more` to succeed. I didn't notice this problem before now
because I let GitHub Actions do that, but now the GitHub CI is no longer
providing a software renderer (or something changed so we stop using it,
but if so, it wasn't the wgpu 0.19 update).

The test can still be run explicitly by naming it, `--test gltf-render`,
which is perfect for the current situation.