mesh: Implement instanced blocks as part of chunk mesh building.
We rebuild a table of instances per-chunk, then copy all visible chunks
to a single table for rendering each frame. This is inefficient in total
CPU work, but it is still an overall win at least for the demo-city,
because it reduces the total mesh data from ~700 MiB to ~50 MiB due to
not duplicating complex blocks.

This commit does not activate instancing but only adds the plumbing that
will make it actually work when activated.

Future work:

* We'd like to be able to update instances from space changes without
  even rebuilding the chunk mesh at all, but right now we don't have any
  code paths to do that; instances are strictly used as a substitute for
  parts of the chunk mesh.
* Smarter instance tables which can be updated incrementally instead of
  having to be thrown out.
* Not strictly related to instancing only, but make sure we don't try to
  render way too many triangles by putting a cap on how many instances
  are ever put in the tables.