mesh: Remove global instance table from `ChunkedSpaceMesh`.
Now the caller can do it on their own, with the help of
`InstanceCollector` if they wish, and avoid both redundant chunk
chart iterations and doing the work in update() instead of draw()
even though it doesn't depend on the Space.

Also updated some documentation to reflect the presence of instancing.