mesh: Analyze the entire block volume before generating vertices.
The premise here is that we can iterate over the full volume once
(instead of six times) and precompute information which lets us skip
parts that we know won't generate any triangles when we do the six
face passes.

Unsurprisingly, in the benchmarks, this results in +25% time for the
checkerboard case and -60% for the fully opaque case, which I consider
a net win.

Additionally, the presence or absence of the block texture is now chosen
before any voxels are computed, so if we wish, we can guarantee to never
mix texturing and vertex colors. (But this is not currently done.)

In the future, this may also enable better mesh generation free of
“T-junctions”, by also computing all the locations where edge or corner
vertices are needed, before generating any triangles.