Adjust polyfill imports to not give warnings when testing w/o `all-is-cubes/std` feature.
It seems that when compiling the lib as a dependency of tests (that is,
with `cfg(test)` *not* set), std is linked anyway and so the trait
imports count as unused because they're shadowed by std inherent
methods. This doesn't quite make sense and is worth further
investigation, but `allow(unused_imports)` will suppress the warnings
that I can't otherwise deal with.