gpu: Fix deadlock in `AtlasAllocator`.
If the last tile handle was dropped while `flush()` was in progress,
then the `TileBacking` would be dropped inside `flush()`, causing a
deadlock on the allocator backing lock. To avoid this, defer the drops.

Another possibility would be to give the `Alloctree` a dedicated mutex,
making it concurrent in itself, since the only thing `TileBacking` is
actually doing is `alloctree.free()`, but I am not sure that wouldn't
complicate resizing in some way.