gpu: Add background meshing tasks.
This seems to introduce a hiccup on desktop (but not wasm) startup,
just when the second frame should be getting rendered, more or less.
I'm not sure what is going on and the self-profiling info doesn't
reveal anything that shouldn't be happening; the `render_data_updater`
calls for blocks are taking a _little_ unreasonable amount of time,
but 250 ms, not multiple seconds, so that's not it.