Rename `URef` to `Handle`, and rename related types.
This rename is in order to avoid confusion between `Handle`s and the
Rust concept of “reference”, as well as “reference counting” in general;
also, `URef` is a bit clunky in itself, and `handle` is a valid variable
name whereas `ref` is not.

I'm not entirely happy with the name `Handle`, but it's short, reads
smoothly, and gives the general impression of “this refers to a value
but cannot simply be dereferenced to obtain it”; and I've wanted to get
rid of the name `URef` for a while now and haven't thought of anything
better.