Improve block tick action conflict processing, and add `Fluff::BlockFault` to report it.
This does not do the “correct” conflict processing, which would detect
exactly which pair(s) of transactions conflict and skip only those;
instead it pauses *all* block ticks, which is still an improvement on
panicking.

A further refinement we could make is to execute only actions whose
bounds are equal to their triggering cube — effectively giving them
higher priority — because those cannot ever conflict with others of that
category. In the long run, however, we want precise detection, which
will require a spatial index data structure to be done efficiently.
(But we have other possible uses for that, too.)