gpu: Remove `GraphicsResourceError` and replace it with `RenderError` or nothing.
There are probably some missing error propagation cases, but this change
doesn't make any of them worse. In particular, drawing functions using
`wgpu` are infallible because, if there are any errors, either they'll
be panics from unhandled wgpu validation errors, which indicate bugs in
our code and thus *should* be panics, or they'll be asynchronous and
therefore not expressible as `Result`s.