desktop: Sketch out gamepad input with `gilrs`.
We will need to teach `InputProcessor` about gamepad-style input,
and a bunch of other things, but this is something we can build on.

Additionally, we want web support, which will require moving this code,
but it's unclear what crate to move it to; `all-is-cubes-ui` is the
logical general place, but I don't like putting platform API bindings in
its dependencies.