Cache light lookups during light evaluation.
This seems to give about ~1% speedup, but I invented it in order to try
to decrease the cost of another change which would make transparent
blocks also care about the light behind them, thus doubling all lookups
without this cache. In that case, the cache seems to offer a 5% speedup.