ui: Move mutable parts of `Session` into new inner struct `Shuttle`.
This will allow the `MainTaskContext` to provide direct mutation access
to the session state rather than indirecting through control messages,
which will allow the main task to have a more consistent view of the
session. However, that is not yet done.

This change should be invisible from the outside; the shuttle is never
unavailable except while the `Session` is exclusively borrowed. In the
future, we might change that to enable the session to perform background
processing of universe steps.