desktop: Fix flaky behavior causing `output-path-with-graphics` test to fail.
I believe the main thread shutting down after printing was racing with
the completion of the PNG export, causing the file to be empty
sometimes. To fix that, we make the `GraphicsType::Print` main loop wait
for the main task.

This is a kludge to fix tests; I think the right solution here is to
rebuild the printing as a form of record/export itself, and eliminate
the dedicated print main loop. However, that will require further work
since the terminal writing is very tied to being a renderer right now.