gpu: Make `LightTexture` sized to the visible area of the `Space`.
This allows rendering spaces that are bigger than the maximum 3D texture
size, at the price of needing to update the texture more often.
Currently, the texture is always tightly fitted to

    min(visible area bounds, space bounds)

In principle it could be larger to perform less frequent updates, but
that seems to me to be a poor tradeoff.