Rerun: use a more general depth unprojection calculation.
This might be slightly more efficient per-pixel, but mainly, it is more
general; it is said to work for infinite-far-plane projections, and I
presume will also work for orthographic, which we might want to do
someday.

We don't strictly need to send the entire inverse projection matrix to
the shader, but doing that is simpler and more obviously-correct than
picking out only the needed elements.