Change `Camera::projection_matrix()` depth range from -1..1 to 0..1.
This is the modern convention (as far as I know, all graphics APIs
other than OpenGL use it) and it simplifies the `wgpu` glue code.

Of course, this results in a bunch of rounding-error changes in the
render tests. But also, it reveals that the raytracer is having
through-the-corner leakage when it really shouldn't (presumably because
the simpler arithmetic allowed the test to luck into a perfectly
diagonal ray). Fixing that is an independent task, so I just adjusted
the tests for now.