mesh: Discover bounding rects during analysis.
This allows the planar meshing phase to only scan the rectangle that
contains visible voxels in this layer, rather than the larger rectangle
sliced from the `Evoxels` bounds.

However, it is a significant additional per-voxel processing cost, so
we need to keep it efficient. It turns out that the best way to do so
that I've found so far is to store entire 3D boxes and not try to do
any `face_transform` rotations. There's probably room for further
improvement here.