transaction: Add `Transactional::transact()`.
This method cleans up the now-common pattern of building a transaction
with mutations and then immediately executing it.

The `Transactional` trait is not really required, but using it avoids
needing to specify the transaction type explicitly, and allows the
Function to be a method of the transaction target, which seems like a
large ergonomic benefit to me.

Ideally, we would have an async version of this function too, but that
is not possible, because the required borrowing pattern is not currently
expressible when writing the future-returning closure it would require.

comments update