Add internal macro `cfg_should_impl_error!`.
This is used for all of our conditional `std::error::Error`
implementations, which will allow the impls to exist when needed to
satisfy `ErrorIfStd` bounds even when the `util` module is split to
another crate and therefore won't necessarily share `cfg`s with the
code that needs to contain the implementations.

All this can hopefully go away some day when `core::error::Error`
exists.