Generate light ray structs as binary rather than text.
I'm planning to switch to generating the entire “propagation table”
at compile time, at which point we'll have a lot more data, which I want
to not have to pass through all of rustc's processing for source code.
So, this commit establishes the new data encoding technique before we
actually change the data.