Reorganize light chart code under one module `space::light::chart`.
(What's with that `mod.rs`, you may be thinking? Well, while I've been
an advocate for the non-`mod.rs` style of module naming, I'm growing
less fond of it because of noticing how it is inconvenient to browse
in any environment that shows one directory at a time. I've decided to
try using `mod.rs`, at least in the cases where a module is composed
nearly entirely of child modules rather than any significant code.)