Lint: Apply `clippy::missing_inline_in_public_items` to `all-is-cubes-base`.
Benchmarking results are unclear, but most of the added `#[inline]`s
should be at worst harmless. My goal with this commit is not to
definitely increase performance but to ensure we aren't burned later by
lack of inlining, and get the lint enabled so we actually think about
all future cases.

(It'd be nice if `missing_inline_in_public_items` ignored all generic
items, but it doesn't.)