Reduce total run time of `space` benchmark; change names.
64×64×64 is so large that it causes the benchmark run times to be
significantly higher. Of course, we'd also like the operations to just
complete fast, but slow benchmarks are a big obstacle to working on
performance, since they require more waiting to do a good comparison.

The renaming is just for more concise output, but in particular, there’s
no reason to have “space” appear once or even twice in all of the names.