Rust 1.79: Apply `unnameable_types` lint.
This catches cases where a type is used in public API but cannot
actually be named. Many of these are intentional (in particular, the
sealed trait pattern, or `doc(hidden)` items where we want to be
minimal about what we expose regardless) but they could also be a
serious mistake.