Add `Transaction::Mismatch` type parameter.
Currently it is always still `PreconditionFailed`, but in the future,
transactions will have individual error types. The purpose of this
feature is to be able to translate transaction failures into in-game
error reports such as highlighting which cubes in a Space are involved,
rather than having only generic and textual information.